replaced mulstiple testfiles with on to obey definitely typed naming
